# openapi.yaml
openapi: 3.1.0
info:
  title: Copilot Session Viewer API
  description: API for viewing and analyzing AI coding assistant sessions from multiple sources (Copilot CLI, Copilot Chat, Claude Code, etc.)
  version: 2.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:3838
    description: Local development
paths:
...
  /api/{source}/sessions/{sessionId}/events:
    get:
      operationId: getSessionEvents
      summary: Get session events
      description: Returns normalized events for a session. Supports optional pagination.
      tags: [sessions]
      parameters:
        - $ref: '#/components/parameters/source'
        - $ref: '#/components/parameters/sessionId'
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
      responses:
        '200':
          description: Session events
          content:
            application/json:
              schema:
                oneOf:
                  - description: Unpaginated — plain array
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                  - description: Paginated response
                    $ref: '#/components/schemas/PaginatedEvents'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/{source}/sessions/{sessionId}/timeline:
    get:
      operationId: getTimeline
      summary: Get timeline data for time analysis
      tags: [sessions]
      parameters:
        - $ref: '#/components/parameters/source'
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Timeline data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Timeline'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/{source}/sessions/{sessionId}/export:
    get:
      operationId: exportSession
      summary: Export session as zip
      tags: [sessions]
      parameters:
        - $ref: '#/components/parameters/source'
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Zip file download
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    source:
      name: source
      in: path
      required: true
      description: 'Session source identifier'
      schema:
        type: string
        enum: [copilot-cli, copilot-chat, claude, modernize, pi-mono]
    sessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-zA-Z0-9_-]+$'
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Event:
      type: object
      required: [type, timestamp]
      properties:
        stableId:
          type: string
        type:
          type: string
          description: 'Event type (e.g. user.message, assistant.message, tool.call, tool.result, subagent.started, subagent.completed)'
        timestamp:
          type: string
          format: date-time
        data:
          type: object
          description: Event-specific payload (varies by type)
    PaginatedEvents:
      type: object
      required: [events, pagination]
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        pagination:
          type: object
          required: [total, limit, offset, hasMore]
          properties:
            total:
              type: integer
            limit:
              type: integer
            offset:
              type: integer
            hasMore:
              type: boolean
    Timeline:
      type: object
      description: Structured timeline data for time analysis visualization
...